refactor(sketches): rename count*/countmin* files to *sketch suffix
- countmin.rs -> countminsketch.rs
- count.rs -> countsketch.rs
- countmin_topk.rs -> countminsketch_topk.rs
- count_topk.rs -> countsketch_topk.rs

Update internal `use` paths and docs accordingly. Add a section in
sketches/mod.rs explaining the *sketch vs *sketch_topk split.

No public type names changed.
